pyproject.toml
[project]
name = "async-tool-calls"
version = "0.1.0"
description = "Experiments with asynchronous, non-blocking tool calls for LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic[vertex]>=0.52.0",
    "google-genai>=1.14.0",
    "pytest>=8.0",
]
